Update use case for Fondant 0.10.1 with lightweight components
