+++ requirements.txt
@@ -1,2 +1,2 @@
-fondant==0.8.0
+fondant==0.10.1
 notebook==7.0.6
