--- src/components/generate_prompts/fondant_component.yaml
@@ -1,13 +1,0 @@
-name: Generate prompts
-description: Component that generates a set of seed prompts
-image: ghcr.io/ml6team/generate_prompts:0.8.0
-
-produces:
-  prompt:
-    type: string
-
-args:
-  n_rows_to_load:
-    description: Optional argument that defines the number of rows to load. Useful for testing pipeline runs on a small scale
-    type: int
-    default: None
