--- src/components/generate_prompts/requirements.txt
@@ -1,1 +1,0 @@
-fondant[aws,azure,gcp, component]==0.8.0
